saas/pyproject.toml
[tool.black]
line-length = 119
target-version = ['py36']
exclude = '''
(
  /(
      \.mypy_cache
    | \.git
    | blueapps
    | blueking
    | migrations
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = 'true'
force_grid_wrap = 0
use_parentheses = 'true'
line_length = 119
skip = [".mypy_cache", ".git", "blueapps", "blueking", "*/migrations"]

[tool.flake8]
ignore = "C901,E203,W503"
max-line-length=119
max-complexity=12
format = "pylint"
show_source = "true"
statistics = "true"
count = "true"
exclude = "*migrations*,*.pyc,.git,__pycache__,node_modules/*,*/templates_module*,*/bin/*,*/settings/*,blueapps,blueking,config,tests/unittest_settings.py"

[tool.mypy]
files=["."]
python_version = 3.6
ignore_missing_imports=true
follow_imports="skip"
strict_optional=true
pretty=true
show_error_codes=true

[[tool.mypy.overrides]]
module = [
    "*.migrations.*",
    "*.blueapps.*",
    "*.blueking.*",
    "*.config.*",
    "*.tests.unittest_settings",
    "*.settings"
]
ignore_errors = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.unittest_settings"
addopts = "--disable-pytest-warnings --reuse-db --nomigrations -s"
python_files = "*_tests.py"
testpaths = [
    "tests"
]

[tool.poetry]
name = "bk_iam"
version = "0.0.0"
description = "IAM（Identity and Access Management） is centralized authority management services  provided by Blueking，which is used for centralized access of various Platforms and SaaS within the BlueKing System, and provides access control for resources that are refined to specific resource instance."
authors = ["TencentBlueKing <contactus_bk@tencent.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "3.6.6"
# blueapps requirement
Django = "2.2.14"
mysqlclient = "2.0.1"
MarkupSafe = "1.0"
Mako = "1.0.6"
requests = "2.24.0"
celery = "3.1.25"
django-celery = "3.2.1"
python-json-logger = "0.1.7"
whitenoise = "5.1.0"
six = "1.15.0"
#v2 requirement
httplib2 = "0.18.1"
# framework
djangorestframework = "3.11.0"
django-filter = "2.3.0"
django-cors-headers = "3.4.0"
Werkzeug = "1.0.1"
drf-yasg = "1.17.1"
dataclasses = "0.7"
pydantic = "1.5.1"
PyJWT = "1.7.1"
cryptography = "2.9.2"
django-prometheus = "2.0.0"
pytz = "2020.1"
aenum = "2.2.6"
# django cache
redis = "3.5.3"
django-redis = "4.12.1"
"dogpile.cache" = "0.9.2"
# preorder tree traversal
django-mptt = "0.11.0"
# profile record
pyinstrument = "3.1.3"

[tool.poetry.dev-dependencies]
# For flake8 support pyproject.toml
toml = "0.10.1"
pyproject-flake8 = "^0.0.1-alpha.2"
flake8-comprehensions = "^3.5.0"
# pytest
pytest = "6.2.2"
pytest-django = "4.1.0"
django-dynamic-fixture = "3.1.1"
converge = "0.9.8"
mock = "1.0.1"
# black
black = "21.7b0"
# mypy
mypy = "0.910"
types-requests = "^2.25.0"
types-pytz = "^2021.1.0"
types-redis = "^3.5.4"
types-six = "^0.1.7"
types-mock = "^0.1.3"
# isort
isort = "^5.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
